mvp/requirements.txt
fastapi==0.115.0 
uvicorn[standard]==0.30.3 
pydantic==2.8.2 
pandas==2.2.2 
numpy==1.24.3 
scikit-learn==1.5.1 
sentence-transformers==3.0.1 
matplotlib==3.9.0 
seaborn==0.13.2 
faiss-cpu==1.8.0.post1 
python-multipart==0.0.9 
